fixing things so they work w/ register_globals off

+++ www/webfiles/mon_devices.php
@@ -13,17 +13,18 @@
 #                                                      #
 ########################################################
 
+require_once("../include/config.php");
 require_once("/var/www/netmrg/lib/stat.php");
 require_once(netmrg_root() . "lib/format.php");
 require_once(netmrg_root() . "lib/auth.php");
 require_once(netmrg_root() . "lib/processing.php");
 check_auth(1);
 
-if ((!isset($action)) || ($action == "doedit") || ($action == "dodelete") || ($action == "doadd") || ($action =="doaddtogrp")) {
+if ((!isset($_REQUEST["action"])) || ($_REQUEST["action"] == "doedit") || ($_REQUEST["action"] == "dodelete") || ($_REQUEST["action"] == "doadd") || ($_REQUEST["action"] =="doaddtogrp")) {
 
-if (($action == "doedit") || ($action == "doadd")) {
+if (!empty($_REQUEST["action"]) && ($_REQUEST["action"] == "doedit" || $_REQUEST["action"] == "doadd")) {
 check_auth(2);
-if ($action == "doedit") {
+if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "doedit") {
 	if ($dev_id == 0) {
 		$db_cmd = "INSERT INTO";
 		$db_end = "";
@@ -54,13 +55,13 @@ if ($action == "doedit") {
 } # done editing
 }
 
-if ($action == "doaddtogrp")
+if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "doaddtogrp")
 {
 	check_auth(2);
 	do_update("INSERT INTO dev_parents SET grp_id=$grp_id, dev_id=$dev_id");
 }
 
-if ($action == "dodelete")
+if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "dodelete")
 {
 	check_auth(2);
 	delete_device($dev_id);
@@ -76,7 +77,7 @@ $group_array = mysql_fetch_array($group_results);
 
 $title = "Monitored Devices in Group '" . $group_array["name"] . "'";
 
-$custom_add_link = "$SCRIPT_NAME?action=add&grp_id=$grp_id";
+$custom_add_link = "{$_SERVER['PHP_SELF']}?action=add&grp_id=$grp_id";
 
 } else {
 
@@ -84,14 +85,14 @@ $title = "Monitored Devices";
 
 }
 begin_page();
-js_confirm_dialog("del", "Are you sure you want to delete device ", " and all associated items?", "$SCRIPT_NAME?action=dodelete&grp_id=$grp_id&dev_id=");
+js_confirm_dialog("del", "Are you sure you want to delete device ", " and all associated items?", "{$_SERVER['PHP_SELF']}?action=dodelete&grp_id={$_REQUEST['grp_id']}&dev_id=");
 make_display_table($title,
-				   "Name", "$SCRIPT_NAME?orderby=name",
+				   "Name", "{$_SERVER['PHP_SELF']}?orderby=name",
 				   "SNMP Options","");
 
 if (!isset($orderby)) { $orderby = "name"; };
 
-if (!(isset($grp_id))) {
+if (!(isset($_REQUEST["grp_id"]))) {
 
 $dev_results = do_query("
 SELECT mon_devices.name AS name, mon_devices.ip, mon_devices.id
@@ -104,7 +105,7 @@ $dev_results = do_query("
 SELECT mon_devices.name AS name, mon_devices.ip, mon_devices.id 
 FROM dev_parents
 LEFT JOIN mon_devices ON dev_parents.dev_id=mon_devices.id 
-WHERE grp_id=$grp_id
+WHERE grp_id={$_REQUEST['grp_id']}
 ORDER BY $orderby");
 
 }
@@ -121,7 +122,7 @@ make_display_item(	$dev_row["name"],"./sub_devices.php?dev_id=$dev_id",
 			formatted_link("View Interface Cache", "snmp_cache_view.php?dev_id=$dev_id") . "&nbsp;" .
 			formatted_link("Recache Interfaces", "recache.php?dev_id=$dev_id") . "&nbsp;" .
 			formatted_link("Recache Disks", "recache.php?dev_id=$dev_id&type=disk"), "",
-			formatted_link("Edit", "$SCRIPT_NAME?action=edit&dev_id=$dev_id") . "&nbsp;" .
+			formatted_link("Edit", "{$_SERVER['PHP_SELF']}?action=edit&dev_id=$dev_id") . "&nbsp;" .
 			formatted_link("Delete", "javascript:del('" . $dev_row["name"] . "', '" . $dev_row["id"] . "')"), "");
 
 } # end devices
@@ -131,20 +132,20 @@ make_display_item(	$dev_row["name"],"./sub_devices.php?dev_id=$dev_id",
 <?
 } # End if no action
 
-if ($action == "add")
+if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "add")
 {
 	check_auth(2);
 	begin_page();
 
 	echo("<big><b>
-	<a href='$SCRIPT_NAME?grp_id=$grp_id&action=addnew'>Create a new device</a><br><br>
-	<a href='$SCRIPT_NAME?grp_id=$grp_id&action=addtogrp>Add an existing device to this group</a>
+	<a href='{$_SERVER['PHP_SELF']}?grp_id={$_REQUEST['grp_id']}&action=addnew'>Create a new device</a><br><br>
+	<a href='{$_SERVER['PHP_SELF']}?grp_id={$_REQUEST['grp_id']}&action=addtogrp>Add an existing device to this group</a>
 	</b></big>");
 
 	end_page();
 }
 
-if ($action == "addtogrp")
+if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "addtogrp")
 {
 
 	check_auth(2);
@@ -152,48 +153,48 @@ if ($action == "addtogrp")
 	make_edit_table("Add Existing Device to a Group");
 	make_edit_select_from_table("Device:","dev_id","mon_devices",-1);
 	make_edit_hidden("action","doaddtogrp");
-	make_edit_hidden("grp_id",$grp_id);
+	make_edit_hidden("grp_id",$_REQUEST["grp_id"]);
 	make_edit_submit_button();
 	make_edit_end();
 	end_page();
 }
 
-if (($action == "edit") || ($action == "addnew")) {
-# Display editing screen
-check_auth(2);
-begin_page();
-if ($action == "addnew") 
-{ 
-	$dev_id = 0; 
-} 
+if (!empty($_REQUEST["action"]) && ($_REQUEST["action"] == "edit" || $_REQUEST["action"] == "addnew")) {
+	# Display editing screen
+	check_auth(2);
+	begin_page();
+	if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "addnew") 
+	{ 
+		$dev_id = 0; 
+	} 
 
-$dev_results = do_query("SELECT * FROM mon_devices WHERE id=$dev_id");
-$dev_row = mysql_fetch_array($dev_results);
-$dev_name = $dev_row["name"];
-$dev_ip = $dev_row["ip"];
-if (!isset($grp_id))
-{
-	$grp_id = $dev_row["group_id"];
-}
-if ($action == "addnew")
-{
-	$dev_row["check_if_number"] = 1;
-}
-make_edit_table("Edit Device");
-make_edit_group("General");
-make_edit_text("Name:","dev_name","25","100",$dev_name);
-make_edit_text("IP or Host Name:","dev_ip","25","100",$dev_ip);
-make_edit_select_from_table("Device Type:","dev_type","mon_device_types",$dev_row["dev_type"]);
-make_edit_checkbox("Disabled (do not monitor this device)","disabled",$dev_row["disabled"]);
-make_edit_group("SNMP");
-make_edit_checkbox("This device uses SNMP", "snmp_enabled", $dev_row["snmp_enabled"]);
-make_edit_text("SNMP Read Community:","snmp_read_community",50,200,$dev_row["snmp_read_community"]);
-make_edit_checkbox("Do not cache interface mappings","snmp_recache",$dev_row["snmp_recache"]);
-make_edit_checkbox("Clear interface cache when interface count changes","snmp_check_ifnumber",$dev_row["snmp_check_ifnumber"]);
-make_edit_hidden("dev_id",$dev_id);
-make_edit_hidden("action","doedit");
-make_edit_submit_button();
-make_edit_end();
+	$dev_results = do_query("SELECT * FROM mon_devices WHERE id={$_REQUEST['dev_id']}");
+	$dev_row = mysql_fetch_array($dev_results);
+	$dev_name = $dev_row["name"];
+	$dev_ip = $dev_row["ip"];
+	if (!isset($_REQUEST["grp_id"]))
+	{
+		$grp_id = $dev_row["group_id"];
+	}
+	if (!empty($_REQUEST["action"]) && $_REQUEST["action"] == "addnew")
+	{
+		$dev_row["check_if_number"] = 1;
+	}
+	make_edit_table("Edit Device");
+	make_edit_group("General");
+	make_edit_text("Name:","dev_name","25","100",$dev_name);
+	make_edit_text("IP or Host Name:","dev_ip","25","100",$dev_ip);
+	make_edit_select_from_table("Device Type:","dev_type","mon_device_types",$dev_row["dev_type"]);
+	make_edit_checkbox("Disabled (do not monitor this device)","disabled",$dev_row["disabled"]);
+	make_edit_group("SNMP");
+	make_edit_checkbox("This device uses SNMP", "snmp_enabled", $dev_row["snmp_enabled"]);
+	make_edit_text("SNMP Read Community:","snmp_read_community",50,200,$dev_row["snmp_read_community"]);
+	make_edit_checkbox("Do not cache interface mappings","snmp_recache",$dev_row["snmp_recache"]);
+	make_edit_checkbox("Clear interface cache when interface count changes","snmp_check_ifnumber",$dev_row["snmp_check_ifnumber"]);
+	make_edit_hidden("dev_id", $_REQUEST["dev_id"]);
+	make_edit_hidden("action","doedit");
+	make_edit_submit_button();
+	make_edit_end();
 
 } # End editing screen
